% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_germplasm_details2.R
\name{ba_germplasm_details2}
\alias{ba_germplasm_details2}
\title{ba_germplasm_details2}
\usage{
ba_germplasm_details2(
  con = NULL,
  germplasmQuery = "",
  rclass = c("tibble", "data.frame", "list", "json")
)
}
\arguments{
\item{con}{list, brapi connection object}

\item{germplasmQuery}{character, the internal database identifier for a
germplasm of which the germplasm details are to be
retrieved e.g. "382"; \strong{REQUIRED ARGUMENT} with
default: "". With this version, can instead be a query eg: '?studyDbId=1001&pageSize=1000" that would follow '/germplasm' in the rest endpoint.}

\item{rclass}{character, class of the object to be returned;  default: "tibble"
, possible other values: "data.frame"/"list"/"json"}
}
\value{
An object of class as defined by rclass containing the germplasm
        details.
}
\description{
Gets germplasm details given an internal database identifier. A LAZY mod of ba_germplasm_details that allows filtering by writing a rest endpoint. Should probably be deprecated at some point.
}
\note{
Tested against: sweetpotatobase, test-server, genesys

BrAPI Version: 1.1, 1.2

BrAPI Status: active
}
\examples{
if (interactive()) {
  library(brapi)

  con <- ba_db()$testserver

  ba_germplasm_details(con, germplasmDbId = "1")
}
}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.2/Specification/Germplasm/Germplasm_GET.md}{github}
}
\seealso{
Other germplasm: 
\code{\link{ba_crosses_study}()},
\code{\link{ba_germplasm_breedingmethods}()},
\code{\link{ba_germplasm_details_study}()},
\code{\link{ba_germplasm_details}()},
\code{\link{ba_germplasm_markerprofiles}()},
\code{\link{ba_germplasm_pedigree}()},
\code{\link{ba_germplasm_progeny}()},
\code{\link{ba_germplasm_search_post}()},
\code{\link{ba_germplasm_search}()},
\code{\link{ba_seedlots_details}()},
\code{\link{ba_studies_germplasm_details}()}

Other brapicore: 
\code{\link{ba_calls}()},
\code{\link{ba_commoncropnames}()},
\code{\link{ba_crops}()},
\code{\link{ba_crosses_study}()},
\code{\link{ba_germplasm_details_study}()},
\code{\link{ba_germplasm_details}()},
\code{\link{ba_germplasm_pedigree}()},
\code{\link{ba_germplasm_progeny}()},
\code{\link{ba_germplasm_search_post}()},
\code{\link{ba_germplasm_search}()},
\code{\link{ba_observationvariables_datatypes}()},
\code{\link{ba_observationvariables_details}()},
\code{\link{ba_observationvariables_ontologies}()},
\code{\link{ba_observationvariables}()},
\code{\link{ba_phenotypes_search_post}()},
\code{\link{ba_programs_search_post}()},
\code{\link{ba_programs}()},
\code{\link{ba_seedlots_details}()},
\code{\link{ba_studies_germplasm_details}()},
\code{\link{ba_traits_details}()},
\code{\link{ba_traits}()}
}
\author{
Reinhard Simon, Maikel Verouden
}
\concept{brapicore}
\concept{germplasm}
